Uplaod all my interesting designs
So Autodesk announced they're going to neuter the "personal" version of
Fusion360 and make it impossible to export designs in STEP/IGES formats.
Hence, I've exported just about everything I've designed that is worth
saving long term and added it here.

This is the precursor to dumping Fusion360 for something else.